+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Currently all releases are considered 'alpha' status, where things may or may no
 
 - Camera/Doorbell support for snapshots and live video re-introduced
 - HomeKit Secure Video recording support may or may not work
+- Should support Nest Thermostat 4th Gen (untested)
 
 ## v0.0.1 (2024-08-27)